refactor: add reusable user search component
